requirements.txt
flask
pdfplumber
markdown
psycopg2-binary
numpy
sentence-transformers
voyageai
